backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Jack Wilburn <jackwilburn@tutanota.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
mysqlclient = "~2.2"
django-cas-ng = "==5.0.1"
djangosaml2 = "^1.12.0"
django = "~5.2"
django-extensions = "^3.2.3"
faker = "^37.0.0"
gunicorn = "^23.0.0"
django-environ = "^0.12.0"
django-cors-headers = "^4.7.0"
